perf(Compiler): Eagerly create bound text nodes as final class fields.
Before, we'd emit something like:

```
class V {
  Text t1;

  build() {
    t1 = Text('');
  }

  detectChanges() {
    if (...) {
      t1.text = newValue;
    }
  }
}
```

After:

```
class V {
  final Text t1 = Text('');

  detectChanges() {
    if (...) {
      t1.text = newValue;
    }
  }
}
```

This makes a large application in Google about 0.17% smaller, as well as makes the compiler and generated output (in Dart) more understandable, so I think it is worth keeping.

In order to compensate for https://github.com/dart-lang/sdk/issues/35381, I also added `createText` as an indirection for `document.createText(...)` to help with code size.

PiperOrigin-RevId: 225221050